server/requirements.txt
openenv-core>=0.2.3
uvicorn>=0.30.0
pydantic>=2.10.0
